testing/CultureVLM_testing/requirements.txt
country_converter
datasets
Pillow
accelerate
numpy<2.0.0
matplotlib
torchvision
tiktoken
einops 
transformers_stream_generator
torch==2.2.2
transformers @ https://github.com/huggingface/transformers/archive/6017f5e8ed33d48096cdf8630d1cc7cbf2550c90.zip
deepseek_vl @ git+https://github.com/deepseek-ai/DeepSeek-VL.git
python-dotenv
google-generativeai
openai